Modifying the Patient.gender mapping example to handle unmapped source code.
Changed the Patient-example.json input to include an invalid and not mapped source code and adjusted the mapping function to handle this. This is for demonstration purposes to show how this can be handled for OMOP to stay compliant with standard codes when needed (i.e. use "0" when not mapped).